+++ stroom-config/stroom-config-global-impl/src/main/java/stroom/config/global/impl/GlobalConfigService.java
@@ -44,7 +44,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-@Singleton // Needs to be singleton to prevent initialise being called multiple times
+@Singleton
+        // Needs to be singleton to prevent initialise being called multiple times
 class GlobalConfigService {
     private static final Logger LOGGER = LoggerFactory.getLogger(GlobalConfigService.class);
     private static final LambdaLogger LAMBDA_LOGGER = LambdaLoggerFactory.getLogger(GlobalConfigService.class);
@@ -175,7 +176,7 @@ class GlobalConfigService {
         // TODO need to run this periodically and cache it, else we have to wait too long
         //   for all nodes to answer
         final DefaultClusterResultCollector<NodeConfigResult> collector = dispatchHelper
-                .execAsync(new NodeConfigClusterTask(securityContext.getUserToken()),
+                .execAsync(new NodeConfigClusterTask(),
                         5,
                         TimeUnit.SECONDS,
                         TargetType.ENABLED);