+++ stroom-core-shared/src/main/java/stroom/task/shared/TaskManagerFields.java
@@ -12,7 +12,6 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 public class TaskManagerFields {
-    public static final String FIELD_SESSION_ID = "Session Id";
     public static final String FIELD_NODE = "Node";
     public static final String FIELD_NAME = "Name";
     public static final String FIELD_USER = "User";
@@ -23,7 +22,6 @@ public class TaskManagerFields {
     private static final List<AbstractField> FIELDS = new ArrayList<>();
     private static final Map<String, AbstractField> FIELD_MAP;
 
-    public static final TextField SESSION_ID = new TextField(FIELD_SESSION_ID);
     public static final TextField NODE = new TextField(FIELD_NODE);
     public static final TextField NAME = new TextField(FIELD_NAME);
     public static final TextField USER = new TextField(FIELD_USER);
@@ -32,7 +30,6 @@ public class TaskManagerFields {
     public static final TextField INFO = new TextField(FIELD_INFO);
 
     static {
-        FIELDS.add(SESSION_ID);
         FIELDS.add(NODE);
         FIELDS.add(NAME);
         FIELDS.add(USER);