+++ stroom-security/stroom-security-api/src/main/java/stroom/security/api/AuthenticationService.java
@@ -19,18 +19,5 @@ package stroom.security.api;
 import stroom.security.shared.User;
 
 public interface AuthenticationService {
-
-    default User getUser(AuthenticationToken token) {
-        if (token == null) {
-            return null;
-        } else {
-            return getUser(token.getUserId());
-        }
-    }
-
-    default User getAdminUser() {
-        return getUser(User.ADMIN_USER_NAME);
-    }
-
     User getUser(final String userId);
 }