+++ stroom-security/stroom-security-impl/src/main/java/stroom/security/impl/AuthenticationConfig.java
@@ -15,9 +15,7 @@ public class AuthenticationConfig implements IsConfig {
     private String authenticationServiceUrl;
     private boolean authenticationRequired = true;
     private boolean verifySsl;
-    private String apiToken = "eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1Mzg2NDM1NTQsInN1YiI6ImFkbWluIiwiaXNzIjoic3Ryb29tIn0.J8dqtQf9gGXQlKU_rAye46lUKlJR8-vcyrYhOD0Rxoc";
     private String authServicesBaseUrl = "http://auth-service:8099";
-    private String durationToWarnBeforeExpiry = "30d";
     private JwtConfig jwtConfig = new JwtConfig();
     private boolean preventLogin;
     private String userNamePattern = "^[a-zA-Z0-9_-]{3,}$";
@@ -60,16 +58,6 @@ public class AuthenticationConfig implements IsConfig {
         this.verifySsl = verifySsl;
     }
 
-    @RequiresRestart(RequiresRestart.RestartScope.UI)
-    @JsonPropertyDescription("The API token Stroom will use to authenticate itself when accessing other services")
-    public String getApiToken() {
-        return apiToken;
-    }
-
-    public void setApiToken(final String apiToken) {
-        this.apiToken = apiToken;
-    }
-
     @JsonPropertyDescription("The URL of the auth service")
     public String getAuthServicesBaseUrl() {
         return authServicesBaseUrl;
@@ -77,15 +65,6 @@ public class AuthenticationConfig implements IsConfig {
 
     public void setAuthServicesBaseUrl(final String authServicesBaseUrl) {
         this.authServicesBaseUrl = authServicesBaseUrl;
-    }
-
-    @JsonProperty("durationToWarnBeforeExpiry")
-    public String getDurationToWarnBeforeExpiry() {
-        return durationToWarnBeforeExpiry;
-    }
-
-    public void setDurationToWarnBeforeExpiry(final String durationToWarnBeforeExpiry) {
-        this.durationToWarnBeforeExpiry = durationToWarnBeforeExpiry;
     }
 
     @JsonProperty("jwt")
@@ -145,9 +124,7 @@ public class AuthenticationConfig implements IsConfig {
         return "AuthenticationConfig{" +
                 "authenticationServiceUrl='" + authenticationServiceUrl + '\'' +
                 ", authenticationRequired=" + authenticationRequired +
-                ", apiToken='" + apiToken + '\'' +
                 ", authServicesBaseUrl='" + authServicesBaseUrl + '\'' +
-                ", durationToWarnBeforeExpiry='" + durationToWarnBeforeExpiry + '\'' +
                 ", preventLogin=" + preventLogin +
                 ", userNamePattern='" + userNamePattern + '\'' +
                 '}';