+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1353** : Removed session identities from tasks.
+
+* Issue **#1352** : Removed internal API token.
+
+* Issue **#1347** : Added `http-call` method to XSLT.
+
 * Change docker image to allow us to pass in the dropwizard command to run, e.g. server|migrate.
 
 * Stop MySQL outputing Note level warnings during migration about things that don't exist when we expect them not to.